requirements.txt
Flask==3.0.0
kubernetes==28.1.0
CloudFlare==2.19.0
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
structlog==23.2.0
colorama==0.4.6
watchdog==3.0.0
PyYAML==6.0.1
jsonschema==4.20.0
retrying==1.3.4 